(import)rdt: support L3 CDP
Support L3 CDP, that is separate settings for L3 code and data paths.

The logic is such that if the configuration contains separate L3 code
and/or L3 data path setting we require that this is also supported by
the system - unless the corresponding 'optional' option(s) in the RDT
configuration has been set to true. The "unified" L3 schema is used as
a fallback value in case CDP is enabled in the system, but, either one
or both of the separate code/data settings is missing.