(import)rdt: make it possible to specify L3 and/or MB optional
Adds a new 'options' section into the RDT configuration, targeted for
common RDT-related config options. Under 'options', the patch specifies
two sub-sections 'l3' and 'mb' which both have one boolean setting, i.e.
'optional'. This can be used to specify that the corresponding schema in
the group configurations is optional, i.e. it is simply ignored if not
supported by the system. This setting defaults to false which results in
an error in case the schema (L3 or MB) is specified in the group
configuration but the technology (L3 or MB allocation) is not available
in the underlying system.