(import)rdt: implement "partitions"
This patch implements a higher-level "partition" hierarchy for RDT
classes that allows easier configuration of exclusive resource
allocation. The partition configuration specifies a portion of the
resources that will be shared by the classes under it.

One conceptual principle is that the resource allocation between
partitions is exclusive (i.e. there are no shared resources between
partitions) when the resource itself allows/supports exclusive
allocation. In practice, when we now have two resources (cache ways and
memory bandwidth), cache allocations are exclusive (no overlap on the
cache ways between partitions), but memory bandwidth may be overlap
(i.e. all partitions may have 100% of bandwidth, for example).

This patch also changes the previously used terminology a bit: we now
use the term 'class' instead of the previously used 'group' or
'resctrl group' in CRI Resource Manager and it's configuration file.