(import)rdt: make l3 bitmasks of classes relative to the partition
Treat (absolute) l3 bitmasks of a class as if it is relative to the
"basemask" of its partition. This makes absolute bitmasks more usable in
the case of multiple partitions where the partition's basemask may
basically be semirandomly allocated anywhere in the available range of
bits.

An example: A partition has an l3 basemask 0xff0 and a class has an
absolute allocation schema 0x3. In this case the true allocated bitmask for
the class (resctrl group in the system level) would be 0x30.