New logger interface
Ditch dependency to the cri-resource-manager logger. Replace it with a
logger interface of our own, imitating the original in order to minimize
changes. Also, specify a wrapper for easily plugging in an
implementation of the standard golang logger interface.

Thus, for application specific logging the users of the library now have
two options: either implement our Logger interface or alternatively plug
in a standard logger implementation using the wrapper. By default, we
use a standard logger baked inside the wrapper.