Get classes from resctrl fs on init
Initialize classes (CTRL and MON groups in resctrl) from the filesystem
on Initialize(). This makes it possible to start using existing setup
even without applying any configuration. This is not perfect, though, as
annotations (that will show as Prometheus metric labels) are not
populated. It is out-of-banc information that is not stored in the
resctrl filesystem.