Preserve order of partitions in relative L3 allocation
Separate calculation of the number of bits allocated/granted from the
formation of the actual L3 schema bitmask. This way we always fill the
schema bitmask in the same partition order when resolving relative L3
allocations. That is, the first partition (partitions are resolved in
alphabetical order) will always get a bitmask starting from bit 0 and
the last partition will get the last bits up to msb.

This makes allocation results more coherent and predictable, and,
ensures that a maximum amount of overlap between code and data path
bitmasks is achieved if separate bitmasks would be specified for a
partition.