Fix bug in cache id parsing
Magically, there were two bugs that happened to counter each other. In
reality, we were not properly parsing the first cache id from L3CODE or
L3DATA schema. However, because the resulting strconv error was ignored
the parsed id was set to 0 which usually happens to be true for the
first cache id.