cgroups: fix building cgroups on non-linux platforms
Cross-platform builds of projects that use the blockio package fails
because the cgroups package uses the NameToHandleAt syscall (not
available on darwin, freebsd), and because Rdev may not be
uint64 (darwin). However the blockio package may be useful on
non-linux hosts when using OCI output. This patch fixes these build
issues and platforms.