go.{mod,sum}: pull in apimachinery v0.21.0.
Pull in the highest released version of apimachinery for
which there is a matching version of *every* other split
out staging package. Otherwise goresctrl can screw up the
dependencies of any package that requires other split out
packages than apimachinery. In the worst case this causes
a dependency set which is inherently incompilable.

The latest such version is v0.21.0 (due to client-go).