rdt: fix default schemata
Fix a bug attempting to write an incorrect (empty) schemata to resctrl
when config of the corresponding control (CAT or MBA) is not specified.
In more detail, an incorrect schemata was generated if CAT was enabled
in the system but no l2/l3allocation was specified in the config (or
similarly MBA was enabled but no mbAllocation was specified in the
config). This patch fixes the error, allocating "100%" (all cache lines
or full memory bandwidth) if nothing is specified in the config. This
aligns with the system default when creating new resctrl groups. At the
same time, this patch ensures that all cache ids are configured into a
known predictable value.