github workflows: split verify in common and per-pr part
Pre-work for further improvements in the CI.

This patch specifies concurrency so that earlier jobs on the same PR get
cancelled (if PR is updated in quick succession). Also, it changes the
workflow to take the Go version from the go.mod file, making future
maintenance easier as there is no more need to edit the github workflow
file when bumping Go version.