sst: add fuzz test for ClosSetup
The test needs to be run as root on an SST-enabled system. Moreover the
test manipulates the actual SST configuration of the target system so a
dedicated non-production system should be used. Test cannot be
parallelized (as it depends on the system state). Because of these
precautions the test needs to be explicitly enabled with
-sst.system-fuzz flag.

An example how to run the test:

    # go test  -parallel=1 -fuzz=FuzzClosSetup  ./pkg/sst/... -sst.system-fuzz