rdt: don't cache control or monitoring groups
Always read from the resctrl filesystem when classes are queried.

The DiscoverClasses() function is dropped and replaced with
SetResctrlGroupPrefix() that alters the prefix used for resctrl group
(directory) names. The functionality is slightly altered, though:

1. Previously the groups belonging to the original prefix were filtered
   out if the new prefix was a substring of the old prefix but this is
   not the case anymore. We don't keep records of the "original" classes
   or prefix - everything matching the new prefix will be returned.
2. Previously the original prefix was still used for monitoring groups.
   This is now changed, the new prefix will be used for monitoring
   groups, too.

Unit tests that needed to be adjusted were modified to use the
github.com/stretchr/testify assert library. Not all tests were converted
in order to limit the number of changed lines. This will be addressed in
follow-up patches.